Fixed Link Shortner Bug
I added a new tag to the release not knowing that it would invalidate the link shortners link so I had to create a new link. BUT my limit on editing the link (backend) from random characters to custom words was Reached for this month. So I had to update the link on the Easy Page to the new link I had created with the link shortener (WITHOUT a custom backend). Which was really unfortunate.